reorganize plugin testing code
plugin testing code is split up from a single file fixtures.py (that
actually did not contain a single fixture) over multiple files

  * one for the manager, which is now called TestManager
    (it has nothing to do with fixtures)
  * one for test classes & runners for use with unittest
  * one for (newly added) fixtures for use with pytest
